Avoid reporting inconsistent events during multi failures
If the instrumentation end up crashing because of a test
failure, it tends to report a second testFailed/testEnded
after the original testStarted/testFailed/testEnded, and
messed up the events order.

In order to be consistent, log the second round of errors
instead of sending proper events.

See bug for details and examples.

Test: unit tests
Bug: 119239432
Change-Id: Ie17cd061af230a198d0fc6bb6218dbe37e0031af