Use R8 for libraries
Add support for R8 usage in libraries. Class file
backend in R8 is ready for preview, and this allows
users that are building AARs to try it out. It is
quite similar to the DEX backend, except for desugaring
that is not possible in the class file backend.

New flag is added, android.enableR8.libraries to turn
on/off usage of R8 in libraries. It is on by default.

BUG: 119550837
Test: R8TransformTest
Change-Id: If031396eaefb3da0d007ee6b25d9f4f2cdc32f18