Fix Swap Hangs
When Task A depensd on Task B and task B fails without someone calling
get(), that task is never detached from the runner.

This happens when swap() fails and cache() never calls get() on it.
We still get a pop-up about the exception, however, the next time
we run, runner.join() just hangs.

Change-Id: I414e2e0003029afb95615368c3b87b8bdcf0c559
Fixes: 120106893
Test: TaskRunnerTest.testPreviousTasksFailed()