Changes the task runner to have explicit run and runAsync [2]
Tasks are not automatically started, they are created and queued.
When run is called, tasks start being executed and the runner
waits until all tasks are done.

There is a new runAsync method that will do this on a separate thread
raising any exception as a runtime exception all the way to the top.

Test: added new tests
Bug: 120506342
Change-Id: Ib05a424ac96bfd4d3a9a1c5d192b51836eb57fdb