Adding AgentStatusSampler.
This sampler looks at the cached status result in the daemon.
The reason we are not sending events from the GenericComponent
directly is because the component only knows about the current
active process. However we should tie agent status to a session.

One alternative way of doing this would be to trigger an event
for all sessions that are tied to the process in the HeartBeat
request. However the request does not determine the state of the
agent it is the GenericComponent thread.

Another way of doing this would require we add a lock
inside GenericComponent that locks sessions from being changed
while we query the agent status for a given pid.

The current approach adds a sampler that looks at the daemon cache
meaning there is a potential of upto a 2 second delay before
studio gets its first event that the agent has attached.

Test: N/A
Bug: 120572690
Change-Id: I9bc843cc337525628b319f8e6fcf4c85be1282c1