Fix compiler working directory.
When the top level CMakeLists.txt references other CMakeLists.txt
through add_subdirectory(), the buildDirectory for each CMake targets are
different from the working directory that Ninja passes to the compiler.
Instead, the buildDirectory for the CMake project seems to be what we
want.

However, note that there still seems to be some bugs related to cache
invalidation after user changes CMakeLists or native code structure.

Bug: https://b/118596839
Test: Updated unit test to cover the basic case. Also manually tested
end-to-end and verify things are working correctly.

Change-Id: I388287d8529caa0748c2c2b97dcd0b8b91ee1147