Fix ApkData serialization.
This is the first part on fixing the serialization of ApkData with the
intent of being able to use it as an @Input on Gradle (or @Nested, as
it's currently the plan).

As we cannot serialize the lazy versionCode and versionName suppliers,
this CL add two specific fields to keep the state when after
deserialization the both transient fields are not available anymore.

Note: *This* serialization is *not* the same as the one done by
BuildOutput/BuildElements flow, which uses the DefaultApkData and
override the suppliers totally.

Bug: 110408030
Test: existing
Change-Id: I6ac824b2b42421e1e5bf5d6bd2f45cd714bdf1f0