Replace SimpleWorkQueue with WorkLimiter
The SimpleWorkQueue would create WorkQueue with threads that would never
be shut down, which would leak the classloader leading to Metaspace
exhaustion.

Replacing SimpleWorkQueue with a new WorkLimiter seems lower risk than
trying to implement thread shutdown in WorkQueue, and as this is the
last user of WorkQueue, this allows WorkQueue to be removed entirely.

Fixes: 123068209
Test: See change I76772c58bea7806bfd976c111c25ac1c0d8acdcb with GradlePluginMemoryLeakTest
Change-Id: I5f56e018d3e07973a163fe164fb0cf1d11bd9863