Fix cacheability of MergeJavaResourcesTask.
Since some of the inputs will contain jar files, we cannot use
@InputFiles which only looks at the jar and not the content.
This means that if entries didn't change but timestamp did(*)
then the task is out of date.

Using @Classpath instead will look at the content of the jars
instead and cache key for the task will be the same.

This CL also adds an empty/.empty entry to BundleLibraryJavaRes
jars to prevent the creation of totally empty jars, which can
confuse gradle's up-to-date check when using the @Classpath
annotation.

(*) this happens because Jar task are not cached so on every run
you can have jars with different timestamp but same content.

Bug: N/A
Change-Id: Icd8a9e5064abbf5ffaf80aff85d0720a0ca72867
Test: N/A