Do not install when APK is unchanged.
Since Delta Install and Swaps are parallel this CL changes
the code path of both.

On the Install side, don't brother with delta install when
all the checksums are identical. Instead, return an empty
string as session ID.

On the Swap side, since we are waiting for the session id,
already, just short circuit everything when we get an empty
string as id.

Bug: 117889823
Test: Swapped / Installed a few times with and without changes.

Change-Id: Ibeb941a7c45ae2a29912019588ecd15696da5c0d