Fix error message in Connection Assistant when no devices
This patch allows the connection assistant to work correctly
when there are no devices connected. Without it, the CA	displays
an error message about an ADB device being undetected by android studio.

The parseAdbLine method was incorrectly treating blank lines as devices
with no name. Note that the only caller already expects this method to
return null for such lines, so no changes are needed outside the parse
method.

Bug: 123886074
Test: AdbDeviceTest
Change-Id: Ib9623caa6814e235005b91e69a1068c9bf2a90eb