Package prebuilt native library only for injected target abi
The existing feature is we only build native code for target ABI when Studio injects target ABI
even though split is not used. However, it does not work well with packaging prebuilt native
libraries and we include all prebuilt native libraries of multiple Abis when user does not use split.

This commit is to let AGP only include prebuilt native library of injected target Abi from Studio
and the new feature is controlled by BUILD_ONLY_TARGET_ABI to be consistent with ExternalNativeBuildTask.

Bug: 124208289
Test: InjectedAbiTest, existing
Change-Id: Ie0479eb156750daa78d4176a9a19cfc360fe94c3