Roll-forward of I16713980f975716ba0dc850a2d53023a9f78dc3d.
(With updated baseline for lint.)

Cap the number of AAPT2 daemons created on MergeResources.

After ag/6556020, when not using WorkerAPI (which is the default) we can
spawn O(CPUs) aapt2 daemons. This CL caps this number to a more
reasonable amount since creating a bunch of aapt2 processes put too much memory
pressure and is also a burden on windows.

I picked 8 since this was the bucketing size used before.

Bug: 125448497
Test: existing + manual verification
Change-Id: Ie389495266f10266c118fff3255fe9835295b9a7