Fix spurious newlines in native build output
The when ever the print function in the constructor is invoked, an
OutputMessage is sent to the gradle daemon client (Android Studio),
which then appends a newline character at the end.

The current implementation of ChunkBytesToLineOutputStream calls print
during flush(), but flush() maybe called regardless of whether there are
newlines in the original input. As a result, each call of flush would
result in an additional newline character in the Android Studio build
window.

Bug: 123776997
Test: manual
Change-Id: Iaaefe8f30f06648c49ab2616b6358abfb35d7c8f