IRR: Enable inheritance install (-p)
Problematic: While developing, the difference between an App featuring
multiple apks (splits), is often contained within a simple split while
others are identical to the previous iteration. Upon installation, it
is a waste of time to feed the very same apk to the Package Manager.

Solution: Upon deltainstalling or delta-preinstalling, leverage the
package manager inheritance flag (-p) which allow to send only apks
which have changed.

The reduced amount of data sent ot the Package Manager is
proportional to the number of untouched splits in the applicaton. On
an app such as Santa Tracker 2018 made of 10 splits, installation is
50% faster and "apply changes" is 40% faster as well.

bug: 127679246
Test: None
Change-Id: I13d2ae70478074ef23fb67bb24620e320f92b5db