Simplify AppPreBuild Task.
Previously the task would do:
- if base module, check for compileOnly AARs.
- check for runtime/compile discrepancy.

compile to match runtime.

Without features, it'll crash at runtime anyway. With features
if might look like it works (if the dependency is added by
the feature) but it's not actually safe because of resource
access.

So this CL removed check #2 and make chekc #1 only present if there
are dynamic features.

We may in the future decide that we don't need this check. It's
expensive and the crash would be pretty obvious at runtime anyway.

Bug: N/A
Change-Id: I8eb4c650af5111e4e09b31a5607876d1c4fba7bc
Test: updated some existing integ tests