Fix MergeJavaResourceTask to depend on javac
Previously, we were adding the output of javac to input of
MergeJavaResourceTask by calling get() on the Provider, but that
stripped away the dependency info.

This CL also changes the BenchmarkTest to run an incremental build in
PSQ, when necessary. Previously, we would only do a clean build and
apply diffs, but this didn't catch issues in the incremental build.

Fixes: 127856561
Test: MergeJavaResourceTaskTest
Change-Id: I2652a3dd83d1b9c7082b21454a1e870d620e9ce2