Add event profiler proto for unified pipeline
Existing Activty/Fragment and System states are moved over
to the interaction_data.proto under the Interaction java
namespace. This avoid having create a redundant set of
enums for the new pipeline.

Cleanups:
-- remove the pid field from the ActivityData and SystemData
proto, since it is not needed beyond the agent->daemon layer.
-- remove the FragmentData message and embed the parent activity
hash directly in ActivityData, since it is only a message containing
a single interger.

Bug: 128992521
Test: Existing

Change-Id: I55b024c2e0637e8a14dd6423a9a0a4fa093c98e8