Fix activity contexts on every AC/ACC deployment
This prevents the contexts from the recreated activities from being out
of date after the new application is installed.

Without this change, if the user does anything that causes a reference
to the old activity's resources to be kept/accessed after the activity
is destroyed (background thread, cache, etc), those resources/contexts
will point to an APK that no longer exists on disk.

Fixes: 127333813
Test: Manual
Change-Id: I28f1f48fa6b21c39afd2c52efc457fb421c8d417