Add a stability exception for support-annotations.
The support-annotations probably have never been backwards
incompatible since the library contains a simple set of
annotations.

This change is actually a workaround a dependency problem in
one of the JetPack libraries: android.arch.lifecycle:common:1.1.1
which has a dependency on the 26.1.0 version of the support
annotations where it probably should have been the 28.0.0 version.

We do not want to display warnings about the support annotations
since it probably is never a real problem.

Bug: 129067099
Test: Unit test added.
Change-Id: I06480a40c2a59ef0cc7dcf7b562e51c1748633b0