Fixed an issue in MergingLog
Fixed an issue in MergingLog where if the entry we are searching for is
the last entry and we are doing a ceiling search, the map search would
return null and we don't check anything. Same thing happens in case of
searching for the first entry and doing a flooring search.
This change fixes these boundary issues.

Bug: N/A
Test: updated MergingLogTest
Change-Id: Ic03798ee0057186215daac31e9c5fc966e433afa