Cleanup error handling in NdkHandler.
This differentiates between the various non-functioning NDK install
states. An NDK that isn't installed is not the same as an NDK
directory that does not validate, and this allows us to store an error
message until we actually know that we need to emit it.

This also cleans up the nullability in NdkPlatform. NdkPlatform
validity is now enforced by the type system so the caller doesn't need
to remember to check isConfigured before using it.

checkNdkDir isn't using this yet, but I have another patch
(https://googleplex-android-review.git.corp.google.com/c/platform/tools/base/+/5017182)
that cleans up that use further that will take advantage of this.
Right now we do not emit useful error messages when that validation
fails (and worse, we do not even perform that validation when using a
side-by-side NDK).

Test: bazel test //tools/base/build-system/...
Bug: None
Change-Id: Id2be673f381442a43a84d8d335cd5b75cd2f71a0