Send API-initiated cpu tracing commands and proto cleanup
- Add StartCpuTrace and StopCpuTrace commands. These are intended
to be shared by all tracing and initiation types.
- Implement logic in the app agent to send these commands to daemon
in the new pipeline. The command processing in the daemon will be
handled in a follow-up CL.
- Refactored CpuProfilerConfiguration into CpuTraceConfiguration, which now
includes the InitiationType and abi info, the previous data is moved into
CpuTraceConfiguration.UserOptions. Also, it is now required to include the
|app_name| parameter in the configuration, because that's what all profiling
types ultimately use. This saves the daemon the extra step of converting
the pid to the app's name, when Studio already have that information.
- Agent is no longer responsible for keeping track of the trace id, since
there should always be a 1-to-1 mapping between start and stop.
- Also fixed a few edge cases where subsequent api-initiated tracing requests
can not start until after the current one is completely processed by
the daemon.

Bug: 119261457
Test: Existing

Change-Id: Ie3295272d768a7dcc6eced92d03609bc7f496f44