Fix regression in merge java resources.
After moving the merge java resources from transforms where streams
were filtered to task where inputs were not, the mergeJavaRes task
started running a lot more often, in particular each time a .class
file would be added/changed/removed in the main project.

Filter the input of the task but retain the original file collection
as the Runnable are not able to deal with fine grained files.

Separated the proguard output (as jar) from the compiler output
(as directories), but we should consider creating two tasks for
the two build path rather than doing inputs branching.

Test: Added integration test.
Bug: 126262668
Change-Id: Ibd3697bc70791718735d76ccfb44fc62aa9e606e