Synchronize access to Android devices.
To interact with Android devices, the AGP provides the DeviceProvider
API. A typical session follows the pattern below:

   deviceProvider.init()
   // Interact with the devices
   ...
   deviceProvider.terminate()

Since multiple sessions can be executed in parallel, we need to
synchronize the entire sessions around a common lock. This CL
implements that in ConnectedDevideProvider, as the abstract
DevideProvider class does not have to promise thread safety.

This CL also adds the DeviceProvider.use(Callable) API, which wraps the
entire session, including calls to init() and terminate(). This is the
new recommended way to use DeviceProvider.

Bug: 128425946
Test: Existing ReadWriteThreadLockTest for the locking mechanism
      + Running ./gradlew connectedAndroidTest on the repro project at
       https://github.com/googlesamples/android-dynamic-features
      (which failed before this CL)

Change-Id: I05b7e6487f8e87fa146e2d4952c40d53684c2b41